[QSA] Bound sparse indices and use 64-bit KV addresses
Clamp sparse attention indices before lookup, keep KV pool address arithmetic
in 64 bits, and guard the degenerate inputs the bounds work exposed: empty
rows, negative and past-the-end indices, and prefill rows whose visible
extent is shorter than the selected budget.

+++ python/sglang/kernels/ops/attention/qsa/mqa.py
@@ -136,9 +136,12 @@ def _qsa_mqa_decode_kernel(
         page_table + batch * stride_pt_batch + page_columns * stride_pt_page,
         mask=valid,
         other=-1,
-    )
+    ).to(tl.int64)
     valid &= pages >= 0
 
+    # int64 page ids: the compressed cache is addressed as
+    # page * page_size * head_dim, which passes 2**31 elements once the pool is
+    # large enough, and Triton keeps the offset in the width of its operands.
     key_ptrs = (
         k_cache
         + pages[:, None] * stride_k_page
